Ryukomik: Address review comments
